backend/repofind/requirements.txt
mysql-connector-python>=8.0.33
GitPython>=3.1.31
python-dotenv>=1.0.0
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
